requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.0

# Biology
biopython>=1.79

# Deep learning
tensorflow>=2.10.0
tf2crf>=0.1.11
tensorflow-addons>=0.19.0

# Visualization
matplotlib>=3.5.0

# Optional but recommended
scikit-learn>=1.0.0
